日付：2024/11/15 　変更者：野口 変更内容：管理者ページに以下を追加。 経歴カテゴリ登録 会社情報変更

+++ app/Http/Controllers/Admin/TagAdminController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tag;
+use App\Models\TalentTag;
+
 use Illuminate\Support\Facades\Session;
 
 class TagAdminController extends Controller
@@ -17,6 +19,13 @@ class TagAdminController extends Controller
 
     public function store(Request $request)
     {
+        $count = Tag::where('TAG_NAME', $request->TAG_NAME)->count();
+        if($count > 0){
+            return redirect()->route('admin')
+            ->with('error', 'タグが既に登録されています。タグ名を変更してください。')
+            ->with('activeTab', 'tag-entry');
+        }
+
         Tag::create($request->all());
 
         if (isset($request->TALENT_ID)) {
@@ -32,6 +41,12 @@ class TagAdminController extends Controller
 
     public function delete($id)
     {
+        $count = TalentTag::where('TAG_ID', $id)->count();
+        if($count > 0){
+            return redirect()->route('admin')
+            ->with('error', 'タグが使用されています。削除できません。')
+            ->with('activeTab', 'tag-entry');
+        }
         Tag::destroy($id);
         return redirect()->route('admin')
         ->with('message', 'タグが削除されました。')
